refactor: mocks 폴더내 auth 파일 이름 변경 (#85)
- data/auth -> data/auth-data
- handlers/auth -> handlers/auth-handlers로 파일 이름 변경
